logger added

+++ Bookapi/views.py
@@ -1,21 +1,95 @@
-from django.shortcuts import render
-from rest_framework import generics
-from .models import *
-from .serializers import *
+import logging
+import datetime
+from rest_framework import generics,status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import filters,status
+from rest_framework import generics, permissions
+from rest_framework.permissions import IsAuthenticated,AllowAny,IsAdminUser
+from rest_framework.generics import *
+from rest_framework import filters
 from django.contrib import messages
 from django.http.response import Http404
 from django.http import Http404
 from django.shortcuts import get_object_or_404
-from rest_framework.generics import *
-from rest_framework import filters
+from django.shortcuts import render
 from django_filters.rest_framework import DjangoFilterBackend
+from django.db.models import Q
+from django.contrib.auth.hashers import check_password
+from .models import *
+from .serializers import *
+from django.contrib.auth.models import Permission
+from django.contrib.contenttypes.models import ContentType
+from rest_framework import status
+
+logger = logging.getLogger(__name__)
+
+class UserRegisterView(generics.CreateAPIView):
+    serializer_class = RegisterSerializers
+    permission_classes = [AllowAny]
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_create(serializer)
+      
+        user = serializer.instance
+        content_type = ContentType.objects.get_for_model(Book)
+        permissions = Permission.objects.filter(
+            codename__in=['create_book','modify_book','destroy_book'],
+            content_type=content_type
+        )
+        user.user_permissions.add(*permissions)
+
+        return Response({"message": "User registered successfully", "data": serializer.data}, status=status.HTTP_201_CREATED)
+
+       
+    def perform_create(self, serializer):
+        serializer.save()
+
+
+class Login(generics.ListAPIView):
+    permission_classes = [AllowAny]
+    serializer_class = LoginSerializer
+
+    def post(self, request):
+        serializer = LoginSerializer(data=request.data)
+        if serializer.is_valid():
+            data = serializer.data
+            try:
+                user_obj = CustomUser.objects.get(
+                    Q(username__iexact=data["username"]) | Q(email__iexact=data["username"])
+                )
+                
+                
+                if user_obj.status == 0:
+                    if check_password(data["password"], user_obj.password):
+                        user_obj.last_login = datetime.datetime.now()
+                        user_obj.save()
+                        refresh_token = RefreshToken.for_user(user_obj)
+                        resp = LoginResponseSerializer(instance=user_obj)
+                        return Response(resp.data, status=status.HTTP_200_OK)
+                    else:
+                        return Response({"message": "Invalid password", "status": "1"})
+                else:
+                    return Response({"message": "you are blocked by admin"})
+
+            except CustomUser.DoesNotExist:
+                logger.error('Invalid user')
+                return Response({"message": "Invalid user", "status": "1"})
+                
+                
+            except Exception as e:
+                logger.error('An error occurred: %s', str(e))
+                return Response({"message": "An error occurred", "status": "1"})
+        else:
+            logger.error('Serializer errors: %s', serializer.errors)
+            return Response(serializer.errors)
 
 class AuthorCreateView(generics.CreateAPIView):
     queryset = Author.objects.all()
     serializer_class = AuthorSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def create(self, request, *args, **kwargs):
         author_name = request.data.get('author_name')
@@ -24,6 +98,7 @@ class AuthorCreateView(generics.CreateAPIView):
     
         if existing_author:
             message = "Author with the name '{}' already exists.".format(author_name)
+            logger.warning(message)
             return Response({"message": message}, status=status.HTTP_400_BAD_REQUEST)
 
         serializer = self.get_serializer(data=request.data)
@@ -38,15 +113,18 @@ class AuthorCreateView(generics.CreateAPIView):
 class AuthorListView(generics.ListAPIView):
     queryset = Author.objects.all()
     serializer_class = AuthorSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
 
 class AuthorDetailView(generics.RetrieveUpdateDestroyAPIView):
     queryset = Author.objects.all()
     serializer_class = AuthorSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
 class GenreCreateView(generics.CreateAPIView):
     queryset = Genre.objects.all()
     serializer_class = GenreSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def create(self, request, *args, **kwargs):
         genre_name = request.data.get('genre_name')
@@ -55,6 +133,7 @@ class GenreCreateView(generics.CreateAPIView):
     
         if existing_genre:
             message = "Genre with the name '{}' already exists.".format(genre_name)
+            logger.warning(message)
             return Response({"message": message}, status=status.HTTP_400_BAD_REQUEST)
 
         serializer = self.get_serializer(data=request.data)
@@ -69,15 +148,18 @@ class GenreCreateView(generics.CreateAPIView):
 class GenreListView(generics.ListAPIView):
     queryset = Genre.objects.all()
     serializer_class = GenreSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
 class GenreDetailCreateView(generics.RetrieveUpdateDestroyAPIView):
     queryset = Genre.objects.all()
     serializer_class = GenreSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
 
 class LibrarianCreateView(generics.CreateAPIView):
     queryset = Librarian.objects.all()
     serializer_class = LibrarianSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def create(self, request, *args, **kwargs):
         librarian_name = request.data.get('librarian_name')
@@ -95,10 +177,12 @@ class LibrarianCreateView(generics.CreateAPIView):
 class LibrarianListView(generics.ListAPIView):
     queryset = Librarian.objects.all()
     serializer_class = LibrarianSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
 class LibrarianRetrieveView(generics.RetrieveAPIView):
     queryset = Librarian.objects.all()
     serializer_class = LibrarianSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def get(self, request, pk, format=None):
         try:
@@ -110,14 +194,17 @@ class LibrarianRetrieveView(generics.RetrieveAPIView):
                 "body": serializer.data,
             }, status=status.HTTP_200_OK)
         except Http404:
+            message = "Librarian with ID '{}' does not exist.".format(pk)
+            logger.error(message)
             return Response({
                 "status": "not found",
-                "message": "Librarian does not exist",
+                "message": message,
             }, status=status.HTTP_404_NOT_FOUND)
 
 class LibrarianUpdateView(generics.UpdateAPIView):
     queryset = Librarian.objects.all()
     serializer_class = LibrarianSerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def put(self, request, pk, format=None):
         instance = self.get_object()
@@ -153,6 +240,7 @@ class LibrarianUpdateView(generics.UpdateAPIView):
 
 class LibrarianDeleteView(generics.DestroyAPIView):
     queryset = Librarian.objects.all()
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def delete(self, request, pk, format=None):
         librarian = self.get_object()
@@ -164,7 +252,8 @@ class LibrarianDeleteView(generics.DestroyAPIView):
                 "Message": "Librarian deleted"
             }, status=204)
         except Exception as e:
-            messages.error(request, f'Failed to delete Librarian. Error: {str(e)}')
+            message = "Failed to delete librarian with ID '{}'. Error: {}".format(pk, str(e))
+            logger.error(message)
             return Response({
                 "Status": "Failed",
                 "Message": "Librarian Deletion Failed"
@@ -174,6 +263,7 @@ class LibrarianDeleteView(generics.DestroyAPIView):
 class LibraryCreateView(generics.CreateAPIView):
     queryset = Library.objects.all()
     serializer_class = LibrarySerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def create(self, request, *args, **kwargs):
         library_title  = request.data.get('library_name')
@@ -190,10 +280,12 @@ class LibraryCreateView(generics.CreateAPIView):
 class LibraryListView(generics.ListAPIView):
     queryset = Library.objects.all()
     serializer_class = LibrarySerializer
+    permission_classes = [IsAuthenticated]
 
 class LibraryRetrieveView(generics.RetrieveAPIView):
     queryset = Library.objects.all()
     serializer_class = LibrarySerializer
+    permission_classes = [IsAuthenticated]
 
     def get(self, request, pk, format=None):
         try:
@@ -205,6 +297,8 @@ class LibraryRetrieveView(generics.RetrieveAPIView):
                 "body": serializer.data,
             }, status=status.HTTP_200_OK)
         except Http404:
+            message = "Library with ID '{}' does not exist.".format(pk)
+            logger.error(message)
             return Response({
                 "status": "not found",
                 "message": "Library does not exist",
@@ -214,6 +308,7 @@ class LibraryRetrieveView(generics.RetrieveAPIView):
 class LibraryUpdateView(generics.UpdateAPIView):
     queryset = Library.objects.all()
     serializer_class = LibrarySerializer
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def put(self, request, pk, format=None):
         instance = self.get_object()
@@ -249,6 +344,7 @@ class LibraryUpdateView(generics.UpdateAPIView):
 
 class LibraryDeleteView(generics.DestroyAPIView):
     queryset = Library.objects.all()
+    permission_classes = [IsAuthenticated, IsAdminUser]
 
     def delete(self, request, pk, format=None):
         library = self.get_object()
@@ -260,7 +356,8 @@ class LibraryDeleteView(generics.DestroyAPIView):
                 "Message": "Library deleted"
             }, status=204)
         except Exception as e:
-            messages.error(request, f'Failed to delete Library. Error: {str(e)}')
+            message = "Failed to delete library with ID '{}'. Error: {}".format(pk, str(e))
+            logger.error(message)
             return Response({
                 "Status": "Failed",
                 "Message": "Library Deletion Failed"
@@ -270,12 +367,17 @@ class LibraryDeleteView(generics.DestroyAPIView):
 class BookCreateView(generics.CreateAPIView):
     queryset = Book.objects.all()
     serializer_class = BookCreateSerializer
+    permission_classes = [IsAuthenticated]
     
  
     def get_serializer_class(self):
         return BookCreateSerializer
  
     def create(self, request, *args, **kwargs):
+        if not request.user.has_perm('Bookapi.create_book'):
+            message = "User '{}' does not have permission to create a book.".format(request.user.username)
+            logger.warning(message)
+            return Response({"message": message}, status=status.HTTP_403_FORBIDDEN)
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
  
@@ -290,9 +392,10 @@ class BookCreateView(generics.CreateAPIView):
 class BookListView(generics.ListAPIView):
     queryset = Book.objects.only('book_name','author','language','price').select_related('author').prefetch_related('genre')
     serializer_class = BookSerializer
+    permission_classes = [IsAuthenticated]
     filter_backends = [filters.SearchFilter]
     search_fields = ['book_name', 'author__author_name', 'genre__genre_name']
-
+    
     def get_queryset(self):
         queryset = super().get_queryset()
 
@@ -305,6 +408,7 @@ class BookListView(generics.ListAPIView):
 class BookRetrieveView(generics.RetrieveAPIView):
     queryset = Book.objects.all()
     serializer_class = BookSerializer
+    permission_classes = [IsAuthenticated]
 
     def get(self, request, pk, format=None):
         try:
@@ -316,17 +420,25 @@ class BookRetrieveView(generics.RetrieveAPIView):
                 "body": serializer.data,
             }, status=status.HTTP_200_OK)
         except Http404:
+            message = "Book with ID '{}' does not exist.".format(pk)
+            logger.error(message)
             return Response({
                 "status": "not found",
-                "message": "Book does not exist",
+                "message": message,
             }, status=status.HTTP_404_NOT_FOUND)
 
 class BookUpdateView(generics.UpdateAPIView):
     queryset = Book.objects.all()
     serializer_class = BookCreateSerializer
-
+    permission_classes = [IsAuthenticated]
     def put(self, request, pk, format=None):
         instance = self.get_object()
+
+        if not request.user.has_perm('Bookapi.modify_book'):
+            message = "User '{}' does not have permission to modify this book.".format(request.user.username)
+            logger.warning(message)
+            return Response({"message": message}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = BookCreateSerializer(instance, data=request.data, partial=True)
 
         if serializer.is_valid():
@@ -343,6 +455,12 @@ class BookUpdateView(generics.UpdateAPIView):
 
     def patch(self, request, pk, format=None):
         instance = self.get_object()
+
+        if not request.user.has_perm('Bookapi.modify_book'):
+            message = "User '{}' does not have permission to modify this book.".format(request.user.username)
+            logger.warning(message)
+            return Response({"message": message}, status=status.HTTP_403_FORBIDDEN)
+
         serializer = BookCreateSerializer(instance, data=request.data, partial=True)
 
         if serializer.is_valid():
@@ -359,9 +477,16 @@ class BookUpdateView(generics.UpdateAPIView):
 
 class BookDeleteView(generics.DestroyAPIView):
     queryset = Book.objects.all()
+    permission_classes = [IsAuthenticated]
 
     def delete(self, request, pk, format=None):
         book = self.get_object()
+
+        if not request.user.has_perm('Bookapi.destroy_book'):
+            message = "User '{}' does not have permission to delete this book.".format(request.user.username)
+            logger.warning(message)
+            return Response({"message": message}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             book.delete()
             messages.success(request, 'Book successfully deleted.')
@@ -370,15 +495,18 @@ class BookDeleteView(generics.DestroyAPIView):
                 "Message": "Book deleted"
             }, status=204)
         except Exception as e:
-            messages.error(request, f'Failed to delete book. Error: {str(e)}')
+            message = "Failed to delete book with ID '{}'. Error: {}".format(pk, str(e))
+            logger.error(message)
+            # messages.error(request, f'Failed to delete book. Error: {str(e)}')
             return Response({
                 "Status": "Failed",
-                "Message": "Book Deletion Failed"
+                "Message": message,
             }, status=500)
 
 
 class BookListByPriceView(generics.ListAPIView):
     serializer_class = BookSerializer
+    permission_classes = [IsAuthenticated]
 
     def get_queryset(self):
         price = self.request.query_params.get('price')
@@ -392,6 +520,7 @@ class BookListByPriceView(generics.ListAPIView):
 
 class BooksByGenreView(ListAPIView):
     serializer_class = BookSerializer
+    permission_classes = [IsAuthenticated]
 
     def get_queryset(self):
